feat: lay out the foundational stuff
