+++ pyproject.toml
@@ -8,6 +8,7 @@ authors = [
 ]
 requires-python = ">=3.10"
 dependencies = [
+    "click>=8.1.8",
     "evdev>=1.7.1 ; sys_platform == 'linux'",
     "keyboard>=0.13.5 ; sys_platform == 'win32'",
     "kisesi>=0.3.0",
@@ -16,7 +17,7 @@ dependencies = [
 ]
 
 [project.scripts]
-mvibes = "mechvibes_lite.__main__:main"
+mvibes = "mechvibes_lite:cli"
 
 [build-system]
 requires = ["hatchling"]
