+++ src/mechvibes_lite/const.py
@@ -3,3 +3,5 @@ import kisesi
 log = kisesi.get_logger(__name__)
 
 APP_NAME = "mechvibes-lite"
+APP_DESCRIPTION = "Mechvibes Lite is an alternative to Mechvibes (it plays sounds when you press keys."
+APP_EPILOG = " * Homepage for Mechvibes Lite is hosted at https://github.com/eeriemyxi/mechvibes-lite "
